requirements.txt
asgiref==3.8.1
Django==5.1.1
django-cors-headers==4.4.0
django-filter==24.3
django-ratelimit==4.1.0
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
djangorestframework-simplejwt==5.3.1
packaging==24.1
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-decouple==3.8
sqlparse==0.5.1
whitenoise==6.7.0
